Update index.d.ts (#203)
Since this parameter doesn't appear to be optional, we should either remove the `The default size is 21.` from the documentation or update the code to default to a size of `21` if it's not passed in. This PR does the former, but the intent is more to prompt a discussion. Also let me know if I'm missing something 😄